- Extracted a common interface called JournalRecord out of TransactionLogRecord. - Added the method readRecords to the JournalInterface.

+++ btm-nio-journal/src/main/java/bitronix/tm/journal/nio/NioLogRecord.java
@@ -143,7 +143,7 @@ class NioLogRecord extends TransactionLogRecord {
         buffer.putInt(getRecordLength());
         buffer.putInt(getHeaderLength());
         buffer.putLong(getTime());
-        buffer.putInt(getSequenceNumber());
+        buffer.putInt((int) getSequenceNumber());
 
         if (buffer.position() != CRC32_POSITION) {
             throw new IllegalStateException("There a bug in the implementation " +
